# openapi(2).json
{"openapi":"3.1.0","info":{"title":"Constellation Simulator","description":"Constellation simulator with effectiveness analytics.\n\n**HTTP**: configs, snapshots, analytics.\n**WebSockets**: live telemetry, routes, control.","version":"0.6.0"},"paths":{"/api/configs":{"get":{"tags":["configs"],"summary":"Api List Configs","operationId":"api_list_configs_api_configs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConfigSummary"},"title":"Response Api List Configs Api Configs Get"}}}}}},"post":{"tags":["configs"],"summary":"Api Save Active","operationId":"api_save_active_api_configs_post","parameters":[{"name":"config_id","in":"query","required":true,"schema":{"type":"string","title":"Config Id"}},{"name":"title","in":"query","required":false,"schema":{"type":"string","default":"","title":"Title"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/configs/{config_id}":{"get":{"tags":["configs"],"summary":"Api Get Config","operationId":"api_get_config_api_configs__config_id__get","parameters":[{"name":"config_id","in":"path","required":true,"schema":{"type":"string","title":"Config Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["configs"],"summary":"Api Delete Config","operationId":"api_delete_config_api_configs__config_id__delete","parameters":[{"name":"config_id","in":"path","required":true,"schema":{"type":"string","title":"Config Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/configs/upload":{"post":{"tags":["configs"],"summary":"Api Upload Config","operationId":"api_upload_config_api_configs_upload_post","parameters":[{"name":"as_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"As Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScenarioModel"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/scenario":{"get":{"tags":["scenario"],"summary":"Api Get Scenario","operationId":"api_get_scenario_api_scenario_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"put":{"tags":["scenario"],"summary":"Api Replace","operationId":"api_replace_api_scenario_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScenarioModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/scenario/load/{config_id}":{"post":{"tags":["scenario"],"summary":"Api Load","operationId":"api_load_api_scenario_load__config_id__post","parameters":[{"name":"config_id","in":"path","required":true,"schema":{"type":"string","title":"Config Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/scenario/new":{"post":{"tags":["scenario"],"summary":"Api New","operationId":"api_new_api_scenario_new_post","parameters":[{"name":"title","in":"query","required":false,"schema":{"type":"string","default":"Untitled","title":"Title"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/scenario/patch":{"post":{"tags":["scenario"],"summary":"Api Patch","operationId":"api_patch_api_scenario_patch_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScenarioPatch"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScenarioAck"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/telemetry":{"get":{"tags":["snapshot"],"summary":"Api Telemetry","operationId":"api_telemetry_api_telemetry_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelemetryFrame"}}}}}}},"/api/routes":{"get":{"tags":["snapshot"],"summary":"Api Routes","operationId":"api_routes_api_routes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutesFrame"}}}}}}},"/api/analytics/coverage":{"get":{"tags":["analytics"],"summary":"Api Coverage","operationId":"api_coverage_api_analytics_coverage_get","parameters":[{"name":"config_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Config Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/vulnerability":{"get":{"tags":["analytics"],"summary":"Api Vulnerability","operationId":"api_vulnerability_api_analytics_vulnerability_get","parameters":[{"name":"config_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Config Id"}},{"name":"top_n","in":"query","required":false,"schema":{"type":"integer","default":12,"title":"Top N"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/satellite/{sid}":{"get":{"tags":["analytics"],"summary":"Api Satellite Impact","operationId":"api_satellite_impact_api_analytics_satellite__sid__get","parameters":[{"name":"sid","in":"path","required":true,"schema":{"type":"string","title":"Sid"}},{"name":"config_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Config Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/compare":{"get":{"tags":["analytics"],"summary":"Api Compare","operationId":"api_compare_api_analytics_compare_get","parameters":[{"name":"a","in":"query","required":true,"schema":{"type":"string","title":"A"}},{"name":"b","in":"query","required":true,"schema":{"type":"string","title":"B"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/export":{"get":{"tags":["analytics"],"summary":"Api Export","operationId":"api_export_api_analytics_export_get","parameters":[{"name":"config_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Config Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/export.json":{"get":{"tags":["analytics"],"summary":"Api Export Json","operationId":"api_export_json_api_analytics_export_json_get","parameters":[{"name":"config_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Config Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ConfigSummary":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"is_example":{"type":"boolean","title":"Is Example","default":false},"satellites":{"type":"integer","title":"Satellites","default":0},"ground_sites":{"type":"integer","title":"Ground Sites","default":0},"updated_at":{"type":"number","title":"Updated At","default":0.0}},"type":"object","required":["id","title"],"title":"ConfigSummary"},"DesignModel":{"properties":{"launch_stage":{"type":"integer","enum":[1,2,3],"title":"Launch Stage","default":3},"planes":{"items":{"$ref":"#/components/schemas/PlaneModel"},"type":"array","title":"Planes","default":[]},"satellites":{"items":{"$ref":"#/components/schemas/SatelliteModel"},"type":"array","title":"Satellites","default":[]}},"type":"object","title":"DesignModel"},"EnvironmentModel":{"properties":{"altitude_km":{"type":"number","maximum":1200.0,"minimum":200.0,"title":"Altitude Km"},"inclination_deg":{"type":"number","maximum":180.0,"exclusiveMinimum":0.0,"title":"Inclination Deg"},"earth_angle0_deg":{"type":"number","title":"Earth Angle0 Deg","default":12.0},"horizon_s":{"type":"integer","maximum":172800.0,"exclusiveMinimum":0.0,"title":"Horizon S"},"step_s":{"type":"integer","exclusiveMinimum":0.0,"title":"Step S"},"min_elevation_deg":{"type":"number","exclusiveMaximum":90.0,"minimum":0.0,"title":"Min Elevation Deg"},"isl_range_km":{"type":"number","maximum":10000.0,"exclusiveMinimum":0.0,"title":"Isl Range Km"},"target_availability":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Target Availability"}},"type":"object","required":["altitude_km","inclination_deg","horizon_s","step_s","min_elevation_deg","isl_range_km","target_availability"],"title":"EnvironmentModel"},"GroundSiteModel":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name","default":""},"role":{"type":"string","enum":["gateway","client"],"title":"Role"},"lat_deg":{"type":"number","maximum":90.0,"minimum":-90.0,"title":"Lat Deg"},"lon_deg":{"type":"number","maximum":180.0,"minimum":-180.0,"title":"Lon Deg"}},"type":"object","required":["id","role","lat_deg","lon_deg"],"title":"GroundSiteModel"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MetaModel":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title","default":""}},"type":"object","required":["id"],"title":"MetaModel"},"PlaneModel":{"properties":{"id":{"type":"string","title":"Id"},"raan_deg":{"type":"number","exclusiveMaximum":360.0,"minimum":0.0,"title":"Raan Deg"},"phase_deg":{"type":"number","exclusiveMaximum":360.0,"minimum":0.0,"title":"Phase Deg"}},"type":"object","required":["id","raan_deg","phase_deg"],"title":"PlaneModel"},"RouteEntry":{"properties":{"gateway":{"type":"string","title":"Gateway"},"client":{"type":"string","title":"Client"},"path":{"items":{"type":"string"},"type":"array","title":"Path"},"hops":{"type":"integer","title":"Hops"},"total_range_km":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Total Range Km"},"propagation_ms":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Propagation Ms"},"reachable":{"type":"boolean","title":"Reachable"}},"type":"object","required":["gateway","client","path","hops","reachable"],"title":"RouteEntry"},"RoutesFrame":{"properties":{"type":{"type":"string","const":"routes","title":"Type","default":"routes"},"t_s":{"type":"number","title":"T S"},"routes":{"items":{"$ref":"#/components/schemas/RouteEntry"},"type":"array","title":"Routes"}},"type":"object","required":["t_s","routes"],"title":"RoutesFrame"},"SatelliteModel":{"properties":{"id":{"type":"string","title":"Id"},"plane_id":{"type":"string","title":"Plane Id"},"slot_deg":{"type":"number","title":"Slot Deg"},"launch_batch":{"type":"integer","enum":[1,2,3],"title":"Launch Batch","default":1}},"type":"object","required":["id","plane_id","slot_deg"],"title":"SatelliteModel"},"SatelliteState":{"properties":{"id":{"type":"string","title":"Id"},"x_km":{"type":"number","title":"X Km"},"y_km":{"type":"number","title":"Y Km"},"z_km":{"type":"number","title":"Z Km"},"active":{"type":"boolean","title":"Active"},"sunlit":{"type":"boolean","title":"Sunlit"}},"type":"object","required":["id","x_km","y_km","z_km","active","sunlit"],"title":"SatelliteState"},"ScenarioAck":{"properties":{"applied":{"additionalProperties":true,"type":"object","title":"Applied"}},"type":"object","required":["applied"],"title":"ScenarioAck"},"ScenarioModel":{"properties":{"schema_version":{"type":"string","const":"cosmo-A-1.0","title":"Schema Version","default":"cosmo-A-1.0"},"meta":{"$ref":"#/components/schemas/MetaModel"},"environment":{"$ref":"#/components/schemas/EnvironmentModel"},"design":{"$ref":"#/components/schemas/DesignModel"},"ground_sites":{"items":{"$ref":"#/components/schemas/GroundSiteModel"},"type":"array","title":"Ground Sites","default":[]},"failures":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Failures","default":[]},"gateway_outages":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Gateway Outages","default":[]}},"type":"object","required":["meta","environment","design"],"title":"ScenarioModel","description":"Full scenario as stored in a config file."},"ScenarioPatch":{"properties":{"move_ground_sites":{"anyOf":[{"items":{"additionalProperties":true,"type":"object"},"type":"array"},{"type":"null"}],"title":"Move Ground Sites"},"add_ground_site":{"anyOf":[{"$ref":"#/components/schemas/GroundSiteModel"},{"type":"null"}]},"remove_ground_site":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Remove Ground Site"},"failures":{"anyOf":[{"items":{"additionalProperties":true,"type":"object"},"type":"array"},{"type":"null"}],"title":"Failures"},"gateway_outages":{"anyOf":[{"items":{"additionalProperties":true,"type":"object"},"type":"array"},{"type":"null"}],"title":"Gateway Outages"},"seek_s":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Seek S"},"step_s":{"anyOf":[{"type":"integer","enum":[30,60,120,300]},{"type":"null"}],"title":"Step S"},"playback":{"anyOf":[{"type":"number","maximum":500.0,"exclusiveMinimum":0.0},{"type":"null"}],"title":"Playback"},"tele_hz":{"anyOf":[{"type":"number","maximum":60.0,"exclusiveMinimum":0.0},{"type":"null"}],"title":"Tele Hz"},"route_hz":{"anyOf":[{"type":"number","maximum":60.0,"exclusiveMinimum":0.0},{"type":"null"}],"title":"Route Hz"}},"type":"object","title":"ScenarioPatch"},"TelemetryFrame":{"properties":{"type":{"type":"string","const":"telemetry","title":"Type","default":"telemetry"},"t_s":{"type":"number","title":"T S"},"horizon_s":{"type":"integer","title":"Horizon S"},"step_s":{"type":"integer","title":"Step S"},"satellites":{"items":{"$ref":"#/components/schemas/SatelliteState"},"type":"array","title":"Satellites"},"edges":{"items":{"items":{},"type":"array"},"type":"array","title":"Edges","default":[]}},"type":"object","required":["t_s","horizon_s","step_s","satellites"],"title":"TelemetryFrame"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}